Revert "avoid props drilling"
This reverts commit 3817c91009e812c51a0f45c7a7957956ad30ea8b.
