+++ packages/files-ui/src/Components/Modules/FileBrowsers/SharedFoldersOverview.tsx
@@ -1,4 +1,4 @@
-import React, { useMemo, useState } from "react"
+import React, { useCallback, useMemo, useState } from "react"
 import {
   Typography,
   Table,
@@ -11,13 +11,14 @@ import {
   Button,
   PlusIcon
 } from "@chainsafe/common-components"
-import { useFiles } from "../../../Contexts/FilesContext"
+import { BucketKeyPermission, useFiles } from "../../../Contexts/FilesContext"
 import { Trans } from "@lingui/macro"
 import { createStyles, makeStyles, useThemeSwitcher } from "@chainsafe/common-theme"
 import { CSFTheme } from "../../../Themes/types"
 import SharedFolderRowWrapper from "./SharedFolderRowWrapper"
 import clsx from "clsx"
 import CreateSharedFolderModal from "./CreateSharedFolderModal"
+import { useFilesApi } from "../../../Contexts/FilesApiContext"
 
 export const desktopSharedGridSettings = "69px 3fr 190px 150px 45px !important"
 export const mobileSharedGridSettings = "3fr 50px 45px !important"
@@ -95,7 +96,8 @@ const useStyles = makeStyles(
 
 const SharedFolderOverview = () => {
   const classes = useStyles()
-  const { buckets, isLoadingBuckets } = useFiles()
+  const { filesApiClient } = useFilesApi()
+  const { buckets, isLoadingBuckets, refreshBuckets } = useFiles()
   const [createSharedFolderModalOpen, setCreateSharedFolderModalOpen] = useState(false)
   const [direction, setDirection] = useState<SortDirection>("ascend")
   const [column, setColumn] = useState<"name" | "size" | "date_uploaded">("name")
@@ -117,6 +119,14 @@ const SharedFolderOverview = () => {
       }
     }
   }
+
+  const handleRename = useCallback((bucket: BucketKeyPermission, newName: string) => {
+    filesApiClient.updateBucket(bucket.id, {
+      ...bucket,
+      name: newName
+    }).then(() => refreshBuckets(false))
+      .catch(console.error)
+  }, [filesApiClient, refreshBuckets])
 
   return (
     <>
@@ -195,6 +205,8 @@ const SharedFolderOverview = () => {
               {bucketsToShow.map((bucket) =>
                 <SharedFolderRowWrapper
                   key={bucket.id}
+                  bucket={bucket}
+                  handleRename={handleRename}
                 />
               )}
             </TableBody>
